fabtests: Add option to disable FI_MR_LOCAL
Add the mr_mode to the set of test options and include as part
of initialization.  Provide a new command line option to disable
specific mode bits to force testing of providers without those
modes enabled.  The only defined option is for FI_MR_LOCAL,
which removes registration from the test apps and checks that
providers handle the buffers correctly.  This allows testing the
MR cache feature in some providers.

Update all tests to reference the new test option.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>