fabtest: Add test to verify MR caches
Introduce a new test that is capable of verifying that MR caches
work correctly.  This test will free (unmap) and allocate (mmap)
data buffers in between each transfer.  It attempts to use the
same virtual address with each mapping, so that the virtual to
physical address changes.  Providers that cache
memory registrations must be able to respond to these changes
without corrupting the memory or crashing the app.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>