prov/rxm: refactor cmap allocation and atomics progress
Start CM thread when enabling the endpoint. This is a revert of

(PR #4251) 645ccf9 prov/rxm: Add support for inserting addresses before EP enabled

The above patch fixes GitHub issue #4190 where one of the entries in
cmap->handles_av table was uninitialized because for any address that was
inserted into the AV before the EP is enabled the corresponding entry in the
handles table wasn't allocated as the cmap itself wasn't allocated.
The above patch fixed it by moving cmap allocation to AV bind. There was still
a gap: no cmap handle update for addresses inserted into AV before the AV was
bound to the endpoint. That issue was fixed by the following patch:

(PR #4502) 1253c8a prov/rxm, util/av: Update CMAP when it is allocated

The second patch fixes both the first and second issue. So moving cmap
allocation to AV bind is not needed anymore.

Furthermore, allocating cmap when the endpoint is enabled allows us to avoid
unnecessary serialization requirements when accessing MSG CQ from atomics
progress thread. So instead of having two separate functions: one that
progresses only MSG EQ until MSG CQ is opened and another that progresses both
MSG EQ and MSG CQ, we need to have only the latter. Opening of MSG CQ is
delayed until user binds RxM CQ just so that we know the wait object requirements.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>