core/mr_cache: Fix possible crash in ofi_mr_cache_cleanup
ofi_mr_cleanup() can be called even if the init call fails.  In
this case, the FI_INFO() debug print will try to access the
cache->domain pointer, which will not be set.  Protect access
to the pointer.

Likewise, the monitor may not be set, so we can't access the
monitor's lock.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>