core/mr_cache: Fix buffer pool initialization
If FI_VERBS_MR_MAX_CACHED_CNT = 0, the mr cache will segfault
trying to grow its internal buffer pool.  The problem can be
traced back to the initilization of the buffer pool.  The
max_cache_cnt value is passed into the buffer as the chunk
size.  It should instead be passed as the max pool size.

Pass max_cache_cnt to the correct argument.  Also provide a
reasonable default chunk count.  Update the buffer pool to
set a default chunk count if one is not given.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>