[v1.7.1x] prov/gni: fix configure.m4 so that GNI
compiles on CLE 7.0.UP01

Replace an incorrect environment variable name in configure.m4
that causes the gni provider to not find the gni_pub.h include
file.  The variable to be corrected is a name in the gni-headers
module.  this was also tested on 6.0.UP07 and it works there
as well

fixes #5005
related to #5006

Signed-off-by: Kevan Rehm <krehm@cray.com>
(cherry picked from commit 192d52ce53e8677878451b8bb3c7ae02d22dc99b)