prov/rxm: replace all cmap->locks with ep->lock and add any missing serialization
cmap is an RxM EP resource. The current code has both ep->lock and cmap->lock
in several places depending on the threading level and progress requested. This
patch removes cmap->lock altogether and uses only ep->lock instead to simplify
serialization of access to cmap and MSG provider objects for auto-progress.

When an app requests auto-progress RxM will promote threading level to
FI_THREAD_SAFE so that ep->lock is not a no-op. This follows the same strategy
employed by:

	3fc3c9a prov/rxm: Allow data FI_PROGRESS_AUTO with FI_ATOMIC capability

But one difference is that we remove passing FI_THREAD_SAFE to the core provider
as RxM EP lock should be sufficient to serialize access to MSG provider resources
in case a restricted threading level was originally requested by the app.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>