prov/cxi: Add fault injection, handle CQ CB faults
Add fault injection framework which can be used to test error paths. This
initial commits supports error injection during internal messaging DMA
and malloc operations. With this, code coverage of cxip_msg.c moves from
~75% to ~80%.

To support faults in these paths, the CQ callback API changed to return
a value. CQ Callbacks return FI_SUCCESS if the event was processed. If
an error occurs in a CQ callback, the underlying EQ is not advanced.

Signed-off-by: Zach Tiffany <ztiffany@cray.com>