Configure.ac: Enable atomics option
Added configure option (--enable-atomics) to control atomics support.
Atomics dependecy appears, if the configuration of OS has libatomic
library, so when we try to build the libfabric, we have libatomic.so
dependecy. In order to exclude all atomics dependencies and compiliers
checkings of one we have to add this very option.

Signed-off-by: Nikita Gusev <nikita.gusev@intel.com>