core/mr_cache: Fix matching for faulting memory regions
When a page fault is reported, we need to locate all regions
that overlap with the faulting region.  This may be different
than how the generic 'find' operation works.  Find searches
for a buffer wholely contained within a region.  Overlapping
regions may be excluded (except in the case that merging
regions is enabled).

To handle this, we need to add new search calls to the
mr cache storage, which trickle down to the underlying
implementation, such as the rbtree.  Those search calls
must match if the faulting region overlaps at all with the
cached region.

As part of this fix, it's possible for an application to
be using the valid portion of a region, while a portion of
it is being freed.  This is permissible, so the use_cnt
assertion ofi_mr_cache_notify is incorrect.  Convert
to an if check, and only free the entry once it is no
longer in use.  Note that the entry is still removed from
the cache, since a portion of the cached region is no
longer valid.

Problem reported by: William Zhang wilzhang@amazon.com

Signed-off-by: Sean Hefty <sean.hefty@intel.com>