src/windows: Change getifaddrs for chosing the right adapter
1) Rewrote geifddars function for using new api - GetAdaptersAddresses.
Why do we use GetAdaptersAddresses?
- get a field -> Ipv4Metric
- add a possibility to support Ipv6 in future

2) Added an algorithm for sorting adapters under Ipv4Metric, using
IP_ADAPTER_ADDRESSES struct.

"Automatic Metric Assignment" assigns a higher metric to a slower
network interface. So, we sort the gotten metrics from the highest
to the slowest one. Then use the first IFF_UP adapter from linked list,
which we formed based on the metric and use for it as a main interface.

Signed-off-by: Nikita Gusev <nikita.gusev@intel.com>