core/iov: Fix range of ofi_iov_end
ofi_iov_end() returns the address of the byte past the
end of the iov, rather than the address of the last byte.
The impact is that two buffers that are adjacent, but do
not overlap, will be reported as overlapping.

This problem shows up in the memory registration cache.

Subtract 1 to return the address of the last byte.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>