prov/verbs: verbs EQ trywait should check the SW queue
Check the software EQ before returning success for EQ trywait. The rdmacm
event channel may still have events but any thread going to wait on the fd should
get woken up. We avoid reading the rdmacm event channel as part of trywait as
we don't know the size of the event that we would receive.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>