prov/rxd: handle errors unpacking packet correctly
If the packet could not be unpacked correctly (due to the packet being smaller than the minimum header side), handle this correctly by returning an error and not processing the packet (treating it like a dropped packet) rather than silently trying to process it.

In the case that retries are disabled, the provider will handle a bad packet by completing in error (since the packet will never be retried). Note that this is ok with smaller transfers, but larger transfers that need data packets to send the rest of the information could have unexpected behavior (since the header/op packet errored out). The better way to handle this would be to track the error state of that peer connection and ignore any data packets from that transfer.

Signed-off-by: aingerson <alexia.ingerson@intel.com>