prov/efa: Enable shared-memory support
Enable shared-memory based transfer with SHM provider for intra-node
communication.

1. For small/medium (adjustable threshold by macro RXR_MEDIUM_SIZE) message,
sender sends RTS packet with real data payload to the receiver via shm
provider. Receiver receives it and extracts the data to the user buffer.

2. For large message, sender only sends the RTS packet with IOV structures
to the receiver. Upon receiving the RTS packet, the receiver first extracts
the IOV structures. Then the receiver issues a RMA READ operation to the
sender's address (included in IOV) to get the real data, which utilizes CMA
copy. Upon receiving the completion of the RMA READ operation, the receiver
sends a EOR packet to notify the sender that data copy is done, so that
sender can write send completion.

Runtime environment variable 'FI_EFA_ENABLE_SHM_TRANSFER' can be used to
enable/disable shared memory based data transfer. It's enabled by default.

This commit needs to bump the version to be protocol compatible.

Signed-off-by: Jie Zhang <zhngaj@amazon.com>