prov/efa: Add peer-aware flow control
The original flow-control protocol was based on the number of posted
receive buffers at an endpoint and a static max credit value that was
computed based on the number of potential peers an endpoint could be
communicating with. While this worked for most cases, we were seeing
inefficient flow-control with single-pair flows with a ping-pong like
communication pattern, where a sender was given fewer credits than what
could have been afforded by a receiving endpoint or that a sender with
multiple flows open to a single peer had enough credits to cause queuing
that could have been avoided.

This patch introduces the following changes:

- Credit management is now at the peer-level, not message-level

- The sender now includes a credit_request along with each RTS. Existing
padding space could be reused, so the RTS size did not grow. This is
derived from the number of credits that it is tracking for the target
peer and based on the number of outstanding tx operations it has to this
peer. A single long message could get increasingly smaller amounts of
credits based on the number of pending sends to the same peer.

- The receiver responds to a credit request taking into account the
default maximum window size for the endpoint, the total number of peers
it is communicating with (from the AV size), the size of the message,
and the actual request itself.

- Once the credit for a long message is negotiated, it is not changed
until the sender runs out of it for that specific long-message exchange
causing another CTS to come in.  Eager messages still do not make use of
the flow-control logic.

- To maintain progress, there is now a minimum number of credits a
sender has to get when the application has a large number of outstanding
sends (configurable, defaulting to 32 based on experimentation).

Signed-off-by: Raghu Raja <craghun@amazon.com>