fabtests/shared: replace ctx arrays with new ft_context structs
The common code only allocates one MR (for both the tx and rx buffers)
and the option to allocate an array of fi_context struct for the
bandwidth tests.

This patch replaces the fi_context arrays with two ft_context arrays
which include an fi_context2 struct for usage. They also include fields
to manage multiple MRs if the test requests for multiple MRs to be
initiated in these structs. If the test does not request multiple MRs,
the array of contexts will be allocated and the buffers will be set to
the common tx_buf/rx_buf.

This requires the following changes:
- Define ft_context struct for each buffer/MR
- Define new mr buffer arrays for each separate buffer
- Change the global tx_cxt_arr and rx_ctx_arr type from fi_context to ft_context
- Remove selective initializiation and cleanup of ctx_arr and replace with
non-conditional allocation of ft_context array of size opts.window_size for
every test
- In the case of the rdm_tagged_peek test, force the window_size to 5
because it previously only allocated an array of contexts of size 5 since
it only sends 5 messages.
- Change existing references of ctx_array to use ft_context::context
- Change receive posting references of tx_cxt_arr to rx_ctx_arr to separate
usage of tx and rx ctx arrays for clarity and future extension for two
way bandwidth tests.
- Add FT_OPT_ALLOC_MULT_MR for allocating opts.window_size buffers/MR regions
- Add FT_TX_MR_KEY to indicate where the keys for the tx MRs will start
- Add FT_RX_MR_KEY to indicate where the keys for the rx MRs will start
- If allocating multiple MRs for the transfers, only allocate the normal
buffer big enough for ctrl messages since they will not be used for the transfers
- Add cleanup functions for all of the regions
- Change send/recv posting function argument from fi_context to void * to
support fi_context2 input

Signed-off-by: aingerson <alexia.ingerson@intel.com>