prov/psm2: Add fi_nic support and use device name for domain name
By default, PSM2 uses all active HFI devices. In order to report fi_nic
information a domain must be associated to a specific HFI device. This
is done by converting a single fi_info structure with default device
selection (-1) into a list of fi_info structures with specific device
settings. Only the PCI bus attributes of the fi_nic information are
populated. The original fi_info structure is kept intact if there are
more than one active devices.

Set the domain name to be the name of the selected HFI device(s).

Update the year in the copyright headers.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>