prov/gni: memcpy() of SMSG data to buffer using incorrect length
Routine __gnix_msg_copy_data_to_recv_addr() was recently changed because
an incoming message larger than the matched recv or multi-recv buffer
would result in buffer overflow. The reverse problem is now occurring,
the recv buffer's req->msg.cum_recv_len is now being used for the length
of the memcpy(), when the incoming SMSG is only a hundred bytes long.

The memcpy() is now called with the minimum of the cum_recv_len and
cum_send_len.

fixes ofiwg#5166

Signed-off-by: Kevan Rehm <krehm@cray.com>