prov/verbs: Remove mr registration macros
Verbs defines macros to generate memory registration functions.
The only difference in the generated functions are whether the
word 'cache' is part of the function name.  The actual generated
code is identical.  We can remove the macros and write the functions
directly to improve maintainability.

As part of the change, we note that verbs only supports registering
a single iov.  So, rather than having the code path call through
to mr_regattr, we can provide the base registration path through
mr_reg.  Since this is the common call invoked by apps, we reduce
the registration overhead by avoiding jumping through unnecessary
calls (regv and regattr).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>