prov/verbs: Save memory descriptor 'entry' when creating md
When a new cache entry is created for a memory region, the
verbs provider receives a callback to mr_cache_add_region.
This allows the provider to register the region with the
underlying verbs device and setup the md.  However, a
reference from the md to the entry is not saved.  That
ends up getting set on return from the cache in the
mr_cache_reg call.  But if the cache was hit, the entry
value has already been saved.  We can avoid extra writes
to update the entry by setting it in mr_cache_add_region.
The resulting code is cleaner in that all md initialization
is done in that call.

While we're updating mr_cache_add_region, remove an unneeded
local variable.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>