include/ofi_mem: set index before calling init_fn
Calling ofi_buf_index in the init_fn was not possible because the
index had not been set. In order to call it, set the region and index
before calling the init_fn.
Also remove the assert in ofi_buf_index to allow the function to be
called without the buffer being used.

Signed-off-by: aingerson <alexia.ingerson@intel.com>