prov/tcp: Bind the specific port range, using SO_REUSEADDR
Issue:
While we try to bind to a port after several unexpected interrupts,
while the app is in process, the port stays in TIME_WAIT status. It is
the specific feature of Linux OS, TIME_WAIT state is what a socket
goes into when it closes. We need to wait, when OS let the specific
port LISTEN again because the app, you try to launch just failed and
cant be launched.
Fixed:
 That is why we should use SO_REUSEADDR(calling tcpx_setup_socket()).
This socket option tells the kernel that even if this port is busy
(in the TIME_WAIT state), go ahead and reuse it anyway. If it is
busy, but with another state, you will still get an address already
in use error.
Tested:
Since ae62148 we haven't use SO_REUSEADDR. We missed TIME_WAIT
problem and let tcpx_setup_socket_nodelay() be without SO_REUSEADDR
and use only TCP_NODELAY.
We've tested now on different port ranges and try to make unexpected
interrupts (Ctr+c) and the app successfully reuse the specific port
even it has TIME_WAIT.

Signed-off-by: Nikita Gusev <nikita.gusev@intel.com>