prov/cxi: multi-receive support
-8 bit rendezvous ID is now used for both long send protocols. For
offloaded rendezvous, the ID is placed in the command. For eager long,
the ID is packed in upper match bits. This simplifies a lot of
rendezvous ID packing code.

-RXC and TXC structures are cleaned up a bit, unused fields removed.

-Accept FI_MULTI_RECV flag in fi_recvmsg(). Multi receive requests set
manage_local in the user buffer LE.

-When a Put event (or any initial event from a transaction) matches a
multi receive request, a child request is created. The child is a
duplicate of the multi-receive request, with fields updated matching the
transaction. Many separate transactions can match a single multi-receive
request, and each is associated with multiple events. The child request
tracks the events for a single transaction associated with the
multi-receive buffer.

-When a child request completes, the request bytes are subtracted from
the parent. If the parent buffer space is exhausted, the child request
completion event will have the FI_MULTI_RECV flag set, indicating the
parent buffer is released.

-When using the eager long protcol, Put events contain reference to the
parent request. A Get is issued using a reference to the child request.

-When using the offloaded rendezvous protocol, all events reference the
parent request, and a unique combination of rendezvous ID and initiator.
This two-tuple is used to look up a child request.

-Reorganized how request fields are populated using events. All requests
have fields filled in recv_req_put_event(), which pulls all necessary
information from the event into the request. This function is now safe
to reenter, simplifying some error code paths.

Signed-off-by: Zach Tiffany <ztiffany@cray.com>