core/mr_cache: Ensure that we remove the correct entry from storage
When we want to remove an cached mr entry from the underlying
rbtree, we use the entry as input into the rbtree find routine.
However, it's possible that we could match some other node
in the tree, depending on which find routine is used.  For
example, find_overlap() can match nodes that refer to an
overlapping memory region.

To avoid possible issues, extend the mr_entry structure to record
the rbtree node pointer.  This will allow us to remove the node
directly without needing to perform a new search.

The new 'storage_context' replaces the 'cached' entry, as it
implies the same state.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>