prov/rxd: Move the receive repost/remove to earlier in the pkt handling
The aim of changes is posting a new recive buffer right away, and don't
wait for finishing the handle of a recive complition. We post a recieve
buffer at once.

The changes have caused the performance part:
The results were run on alltoall test, and it shows us
better results from 64Kb to 4Mb with comparing 'without the current
changes'. But see a little regress from 1b to 64Kb, the diff is
about ~0.08 usec (not bigger).

Signed-off-by: Nikita Gusev <nikita.gusev@intel.com>