prov/rxd: Report all failures when inserting addresses into an AV
rxd_av_insert_dg_addr() calls ofi_rbmap_insert() to insert the
address of the peer datagram endpoint into the local AV.  It
ignores failures if the address is already in the rbtree
(-FI_EALREADY).  However, ofi_rbmap_find() is called prior to
calling rxd_av_insert_dg_addr().  So the address should not be
in the map.

The rxd_av_remove() path assumes that an address is only in
the rbmap exactly once.  So, we can't ignore FI_EALREADY
failures when inserting an address into the rbtree.  We
can't return to the user two different fi_addr values that
reference the same datagram endpoint address, or the remove
call will not work.  (We would need referencing counting
to handle this).

So... rxd_av_insert_dg_addr() is ignoring a failure that should
not occur, because duplicate insertions are checking before trying
to insert the address.  And even if that were not the case, the
rxd_av_remove() call would not work if 2 fi_addr values were
associated with the same EP address.

This is a long winded way of saying that the FI_EALREADY
check can be turned into an assert().

Signed-off-by: Sean Hefty <sean.hefty@intel.com>