Added initialization for ranks for each process.
For more advanced and different testing patterns, the ability for
each process to know its own rank is required.

Users also will use ranks on processes for other collective communication
patterns.

Setup is done by having the server send off a rank to each other process
in the order that they are listed in the pm_job.clients table.