prov/verbs: Simplify XRC physical INI QP connection rescheduling
Move the determination of whether a XRC physical QP connection that
did not complete and that a pending endpoint should be scheduled into
the release of the endpoints INI connection object reference. This
handles cases where an EP that has issued a physical XRC QP connect
and is closed before the connection has been accepted/rejected.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>