prov/efa: Fix mismanagement of pinned memory after fork()
Linux copy-on-write semantics mean that following a fork() call,
parent and child processes will have page table entries pointing to
the same physical page. Since these pages are write protected, if
either the parent or child writes to the page, the hardware will
trap the event. The kernel will then allocate a new page and copy
the contents from the original page, breaking the virtual to physical
page link for that process.

To prevent this case, marking pinned memory with MADV_DONTFORK
only allows the memory range to be seen by the parent process, and
the copy-on-write semantics no longer apply to this memory range.

Since the rdma-core library already does all this work for us,
when we add rdma-core, we will move the logic down a layer and
remove it from here.

Signed-off-by: William Zhang <wilzhang@amazon.com>