prov/efa : fix a potential memory leak in rxr_tx()
Currently in function rxr_tx(), if a RTS packet was not sent
successfully and the error code is not -FI_EAGAIN, the function
will not release the tx_entry allocated in this function,
causing a memory leak.

This patch fix this by adding a call to rxr_release_tx_entry() to
that code path.

Signed-off-by: Wei Zhang <wzam@amazon.com>