Address Vector suppport for FI_SOCKADDR
If the domain address format is set to FI_SOCKADDR then address vector
will be initialized for ipv6 addresses.  If ipv4 addresses are inserted
into the address vector this will cause lookup errors and duplicate
entries and possibly duplicate connections.  The address length needs
to be taken into account with HASH_ADD and HASH_FIND instead of relying
on the length of the address vector. Also zero out the extra bits when
inserting an ipv4 address as a precautionary measure.

Signed-off-by: Matt Wycklendt <wycklendt@google.com>