prov/gni: fix memory leak in multi-recv buffer handling
Routine __handle_mrecv_req() checks if a buffer is a multi-recv
buffer, and if so, it allocates a separate gnix_fab_req struct to
use for the message about to be landed in that buffer.  However,
code in __smsg_eager_msg_w_data() and __smsg_rndzv_start() would
interpret that returned req pointer as a pointer to the multi-recv
buffer, and would not free the 'req' after completing the message
transfer.  The result was the leaking of one gnix_fab_req struct
for each message landing in a multi-recv buffer.

This PR removes the 'multi_recv' variable handling in those two
routines so that the gnix_fab_req struct that gets allocated by
__handle_mrecv_req gets properly freed in the same way that a
non-multi-recv gnix_fab_req would be freed.

fixes #5232

Signed-off-by: Kevan rehm <krehm@cray.com>