core/mr_cache: Limit cache size by default
The MR cache is currently unlimited in how much memory may be
registered by the cache.  At scale, this causes application
crashes when the system runs out of memory.  Change the default
to limit the cache size to:

total memory / # of cpus / 2

Assuming 1 process per core, this will limit the total amount
of system memory that will be maintained by the collective
caches to half of system memory.

Disable the cache if cache size is set to 0.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>