mr/cache: Add non-cacheable registration call
Allow inserting some regions into a cache,
while explicitly keeping others from being added.
The MR cache already provides tracking for this.  We
just need a way to specify that a region should not be
inserted.

Add a new call, ofi_mr_cache_reg() that will register
the given memory region without attempting to insert it
into the cache.  We add a new call, so that multi-threaded
apps that go through this path are not heavily serialized.

The new registration call uses the same data structure for
internal tracking purposes, but otherwise leaves the cache
alone.  I.e. it does not flush entries or change existing
registration calls.

Callers will hook into this call to register memory that
should not be cached.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>